==> jupyter_notebook/01_ETL.ipynb <==
# ETL

## Librerías a trabajar 

import pandas as pd
import numpy as np
import ast

# Explorando el dataset

# Edite 03 filas del CSV por salto de linea 
# fila 19764
# fila 29573
# fila 35672

#  Cargar el dataset
df = pd.read_csv("1_datasets_origen/movies_dataset.csv")
df.shape

df

df.info()

df.describe()

## Transformaciones solicitas

Transformaciones: Para este MVP no necesitas perfección, ¡necesitas rapidez! 
⏩ Vas a hacer estas, y solo estas, transformaciones a los datos:

* Algunos campos, como belongs_to_collection, production_companies y otros (ver diccionario de datos) están anidados, esto es o bien tienen un diccionario o una lista como valores en cada fila, ¡deberán desanidarlos para poder y unirlos al dataset de nuevo hacer alguna de las consultas de la API! O bien buscar la manera de acceder a esos datos sin desanidarlos.

# Crear una lista para almacenar los datos de los campos: 
# belongs_to_collection, genres, production_companies, production_countries, spoken_languages

collection_data = []
genres_data = []
companies_data = []
countries_data = []
spoken_data = []

# Recorrer las filas del DataFrame, para desanidar los campos
for index, row in df.iterrows():
    # Obtener el valor del campo como un diccionario
    collection_dict_str = row['belongs_to_collection']
    genres_dict_str = row['genres']    
    companies_dict_str = row['production_companies']
    countries_dict_str = row['production_countries']
    spoken_dict_str = row['spoken_languages']


    collection_dict = ast.literal_eval(collection_dict_str) if pd.notna(collection_dict_str) else None
    if collection_dict is not None:               
        # Agregar el ID al diccionario con la clave 'id_movie'        
        collection_dict['id_movie'] = row['id']
        # Agregar el diccionario a la lista
        collection_data.append(collection_dict)
    
    genres_list = ast.literal_eval(genres_dict_str)
    for genres_diccionario in genres_list:
        if genres_diccionario is not None:               
            # Agregar el ID al diccionario con la clave 'id_movie'        
            genres_diccionario['id_movie'] = row['id']
            # Agregar el diccionario a la lista
            genres_data.append(genres_diccionario)

    companies_list = ast.literal_eval(companies_dict_str)
    for companies_diccionario in companies_list:
        if companies_diccionario is not None:               
            # Agregar el ID al diccionario con la clave 'id_movie'        
            companies_diccionario['id_movie'] = row['id']
            # Agregar el diccionario a la lista
            companies_data.append(companies_diccionario)

    countries_list = ast.literal_eval(countries_dict_str)
    for countries_diccionario in countries_list:
        if countries_diccionario is not None:               
            # Agregar el ID al diccionario con la clave 'id_movie'        
            countries_diccionario['id_movie'] = row['id']
            # Agregar el diccionario a la lista
            countries_data.append(countries_diccionario)

    spoken_list = ast.literal_eval(spoken_dict_str)
    for spoken_diccionario in spoken_list:
        if spoken_diccionario is not None:               
            # Agregar el ID al diccionario con la clave 'id_movie'        
            spoken_diccionario['id_movie'] = row['id']
            # Agregar el diccionario a la lista
            spoken_data.append(spoken_diccionario)


# Crear un nuevo DataFrame con los datos de la lista
collection_df = pd.DataFrame(collection_data)

# Cambiar los nombres de las columnas
collection_df.rename(columns={'id': 'id_collection', 'name': 'name_collection'}, inplace=True)

# Ordenar las columnas
collection_df = collection_df[['id_movie', 'id_collection', 'name_collection', 'poster_path', 'backdrop_path']]

collection_df


# Crear un nuevo DataFrame con los datos de la lista
genres_def = pd.DataFrame(genres_data)

# Cambiar los nombres de las columnas
genres_def.rename(columns={'id': 'id_genres', 'name': 'name_genres'}, inplace=True)

# Ordenar las columnas
genres_def = genres_def[['id_movie', 'id_genres', 'name_genres']]

genres_def


# Crear un nuevo DataFrame con los datos de la lista
companies_def = pd.DataFrame(companies_data)

# Cambiar los nombres de las columnas
companies_def.rename(columns={'id': 'id_companies', 'name': 'name_companies'}, inplace=True)

# Ordenar las columnas
companies_def = companies_def[['id_movie', 'id_companies', 'name_companies']]

companies_def

# Crear un nuevo DataFrame con los datos de la lista
countries_def = pd.DataFrame(countries_data)

# Cambiar los nombres de las columnas
countries_def.rename(columns={'name': 'name_countries'}, inplace=True)

# Ordenar las columnas
countries_def = countries_def[['id_movie', 'iso_3166_1', 'name_countries']]

countries_def

# Crear un nuevo DataFrame con los datos de la lista
spoken_def = pd.DataFrame(spoken_data)

# Cambiar los nombres de las columnas
spoken_def.rename(columns={'name': 'name_spoken'}, inplace=True)

# Ordenar las columnas
spoken_def = spoken_def[['id_movie', 'iso_639_1', 'name_spoken']]


spoken_def

# Guardar los DataFrames en un archivo CSV
collection_df.to_csv("2_datasets_etl/collection_data.csv", index=False)
genres_def.to_csv("2_datasets_etl/genres_data.csv", index=False)
companies_def.to_csv("2_datasets_etl/companies_data.csv", index=False)
countries_def.to_csv("2_datasets_etl/countries_data.csv", index=False)
spoken_def.to_csv("2_datasets_etl/spoken_data.csv", index=False)

# Se tiene la información en archivos diferentes ya no hace falta tenerlo en el archivo principal 
# Lista con los nombres de las columnas a eliminar
columnas_a_eliminar = ['belongs_to_collection', 'genres', 'production_companies', 'production_countries', 'spoken_languages']

# Eliminar las columnas del DataFrame
df = df.drop(columns=columnas_a_eliminar)

# Mostrar el DataFrame resultante
# nos quedamos con 24 columnas 
df

* Los valores nulos de los campos revenue, budget deben ser rellenados por el número 0.

# Contar la cantidad de valores nulos en las columnas 
print(df['revenue'].isnull().sum())
print(df['budget'].isnull().sum())

# Rellenar los valores nulos de las columnas con 0
df['revenue'].fillna(0, inplace=True)
df['budget'].fillna(0, inplace=True)

* Crear la columna con el retorno de inversión, llamada return con los campos revenue y budget, dividiendo estas dos últimas revenue / budget, cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0.

df['return'] = 0
df['return'] = df['revenue'] / df['budget']
df['return'].fillna(0, inplace=True)
df['return'].replace([np.inf, -np.inf], 0, inplace=True)

df

* Los valores nulos del campo release date deben eliminarse.

# Contar la cantidad de valores nulos en las columnas 
print(df['release_date'].isnull().sum())

# Eliminar filas con valores nulos en la columna 'release_date'
df = df.dropna(subset=['release_date'])

df

* De haber fechas, deberán tener el formato AAAA-mm-dd, además deberán crear la columna release_year donde extraerán el año de la fecha de estreno.

# Convertir 'release_date' a formato de fecha (AAAA-mm-dd)
#df['release_date'] = pd.to_datetime(df['release_date'], format='%Y-%m-%d', errors='coerce')
#df.loc[:, 'release_date'] = pd.to_datetime(df['release_date'], format='%Y-%m-%d', errors='coerce')
df['release_date'] = pd.to_datetime(df['release_date'], format='%Y-%m-%d', errors='coerce')  # Convertir a objeto de fecha y hora


# Crear la columna 'release_year' con el año de 'release_date'
# df['release_year'] = df['release_date'].dt.year
# df.loc[:, 'release_year'] = df['release_date'].dt.year
df['release_year'] = df['release_date'].dt.year


* Eliminar las columnas que no serán utilizadas, video,imdb_id,adult,original_title,poster_path y homepage.

# Eliminar columnas 
df = df.drop(columns=['video','imdb_id','adult','original_title','poster_path', 'homepage'])

Continuar Limpieza

# Obtener las columnas con valores nulos o vacíos
columnas_con_nulos = df.columns[df.isnull().any()]

# Mostrar los nombres de las columnas con valores nulos o vacíos
print(columnas_con_nulos)
print(df.isna().sum())


# Tratar datos duplicados
df = df.drop_duplicates()

#  Identificar columnas con los mismos valores
duplicated_columns = []
for col1 in df.columns:
    for col2 in df.columns:
        if col1 != col2 and df[col1].equals(df[col2]):
            duplicated_columns.append(col2)

#  Eliminar las columnas duplicadas
df.drop(columns=duplicated_columns, inplace=True)

duplicated_columns

# Paso 11: Guardar el DataFrame limpio
df.to_csv("2_datasets_etl/movies_dataset_cleaned.csv", index=False)


# Ahora trabajamos el archivo credits 

df_credit = pd.read_csv("1_datasets_origen/credits.csv")
df_credit.shape

df_credit

# Crear una lista para almacenar los datos de los campos "cast" y "crew"
cast_data = []
crew_data = []

# Recorrer las filas del DataFrame
for index, row in df_credit.iterrows():
    # Obtener el valor del campo como un diccionario
    cast_dict_str = row['cast']
    crew_dict_str = row['crew']    

    cast_list = ast.literal_eval(cast_dict_str)
    for cast_diccionario in cast_list:
        if cast_diccionario is not None:               
            # Agregar el ID al diccionario con la clave 'id_movie'        
            cast_diccionario['id_movie'] = row['id']
            # Agregar el diccionario a la lista
            cast_data.append(cast_diccionario)

    crew_list = ast.literal_eval(crew_dict_str)
    for crew_diccionario in crew_list:
        if crew_diccionario is not None:               
            # Agregar el ID al diccionario con la clave 'id_movie'        
            crew_diccionario['id_movie'] = row['id']
            # Agregar el diccionario a la lista
            crew_data.append(crew_diccionario)

# Crear un nuevo DataFrame con los datos de la lista
cast_df = pd.DataFrame(cast_data)

# Cambiar los nombres de las columnas
cast_df.rename(columns={'id': 'id_cast', 'name': 'name_cast'}, inplace=True)

# Ordenar las columnas
cast_df = cast_df[['id_movie', 'cast_id','character','credit_id','gender','id_cast','name_cast','order','profile_path']]

cast_df

# Crear un nuevo DataFrame con los datos de la lista
crew_df = pd.DataFrame(crew_data)

# Cambiar los nombres de las columnas
crew_df.rename(columns={'id': 'id_crew', 'name': 'name_crew'}, inplace=True)

# Ordenar las columnas
crew_df = crew_df[['id_movie','credit_id','department','gender','id_crew','job','name_crew','profile_path']]

crew_df
	

# Guardar los DataFrames en un archivo CSV
cast_df.to_csv("2_datasets_etl/cast_data.csv", index=False)
crew_df.to_csv("2_datasets_etl/crew_data.csv", index=False)

# Mostrar información básica de cada DataFrame
print("Información básica del DataFrame 'movies_df':")
print(df.info())
print("\nInformación básica del DataFrame 'countries_def':")
print(countries_def.info())
print("\nInformación básica del DataFrame 'spoken_df':")
print(spoken_def.info())
print("\nInformación básica del DataFrame 'spoken_def':")
print(collection_df.info())
print("\nInformación básica del DataFrame 'companies_def':")
print(companies_def.info())
print("\nInformación básica del DataFrame 'genres_def':")
print(genres_def.info())
print("\nInformación básica del DataFrame 'cast_df':")
print(cast_df.info())
print("\nInformación básica del DataFrame 'crew_df':")
print(crew_df.info())
